+++ src/BenchProgram.h
@@ -147,7 +147,7 @@ private:
 
     void Init(const std::string &workDirPath, bool no_clean_up);
 
-    bool buildFull(const std::string &subDir, time_t timeout_limit = 0, bool force_reconf = false,std::vector<int> compile_macro=std::vector<int>());
+    bool buildFull(const std::string &subDir, time_t timeout_limit = 0, bool force_reconf = false,std::vector<long long> compile_macro=std::vector<long long>());
 
     void getCompileMisc(const std::string &src_file, std::string &build_dir, std::vector<std::string> &build_args);
 
@@ -193,10 +193,10 @@ public:
             const std::string &code);
 
     bool buildSubDir(const std::string &subDir, const std::string &wrapScript,
-            const EnvMapTy &envMap,std::vector<int> compile_macro=std::vector<int>());
+            const EnvMapTy &envMap,std::vector<long long> compile_macro=std::vector<long long>());
 
     bool buildWithRepairedCode(const std::string &wrapScript, const EnvMapTy &envMap,
-            const std::map<std::string, std::string> &fileCodeMap,std::vector<int> compile_macro=std::vector<int>(),bool createFile=false);
+            const std::map<std::string, std::string> &fileCodeMap,std::vector<long long> compile_macro=std::vector<long long>(),bool createFile=false);
 
     TestCaseSetTy testSet(const std::string &subDir, const TestCaseSetTy &case_set,
             const EnvMapTy &envMap, bool pass_basic_src_dir = false);
